requirements.txt
# SPS AI Service - Requirements
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
anthropic>=0.18,<1.0
openai>=1.12,<2.0
sentence-transformers>=2.5,<3.0
numpy>=1.26,<2.0
pypdf>=4.0,<5.0
python-docx>=1.1,<2.0
faiss-cpu>=1.7,<2.0
requests>=2.31,<3.0
tiktoken>=0.6,<1.0
